refactor(scales): move scales to /utils/scales
