+++ src/lib/series/legend.test.ts
@@ -1,5 +1,5 @@
 import { AxisId, getAxisId, getGroupId, getSpecId, SpecId } from '../../utils/ids';
-import { ScaleType } from '../utils/scales/scales';
+import { ScaleType } from '../../utils/scales/scales';
 import { computeLegend, getSeriesColorLabel } from './legend';
 import { DataSeriesColorsValues } from './series';
 import { AxisSpec, BasicSeriesSpec, Position } from './specs';
