+++ src/lib/series/rendering.ts
@@ -10,8 +10,8 @@ import {
   BarSeriesStyle,
 } from '../themes/theme';
 import { SpecId } from '../../utils/ids';
-import { isLogarithmicScale } from '../utils/scales/scale_continuous';
-import { Scale, ScaleType } from '../utils/scales/scales';
+import { isLogarithmicScale } from '../../utils/scales/scale_continuous';
+import { Scale, ScaleType } from '../../utils/scales/scales';
 import { CurveType, getCurveFactory } from '../../utils/curves';
 import { LegendItem } from './legend';
 import { DataSeriesDatum } from './series';
