+++ src/lib/series/scales.test.ts
@@ -1,5 +1,5 @@
 import { getGroupId } from '../../utils/ids';
-import { ScaleType } from '../utils/scales/scales';
+import { ScaleType } from '../../utils/scales/scales';
 import { XDomain } from './domains/x_domain';
 import { computeXScale, countBarsInCluster } from './scales';
 import { FormattedDataSeries } from './series';
