+++ src/lib/series/tooltip.test.ts
@@ -1,5 +1,5 @@
 import { getAxisId, getGroupId, getSpecId } from '../../utils/ids';
-import { ScaleType } from '../utils/scales/scales';
+import { ScaleType } from '../../utils/scales/scales';
 import { BarGeometry } from './rendering';
 import { AxisSpec, BarSeriesSpec, Position } from './specs';
 import { formatTooltip } from './tooltip';
