+++ src/state/chart_state.timescales.test.ts
@@ -3,7 +3,7 @@ import { LineSeriesSpec } from '../lib/series/specs';
 import { LIGHT_THEME } from '../lib/themes/light_theme';
 import { mergeWithDefaultTheme } from '../lib/themes/theme';
 import { getGroupId, getSpecId } from '../utils/ids';
-import { ScaleType } from '../lib/utils/scales/scales';
+import { ScaleType } from '../utils/scales/scales';
 import { ChartStore } from './chart_state';
 
 describe('Render chart', () => {
