+++ src/state/chart_state.ts
@@ -61,7 +61,7 @@ import {
   TooltipValue,
   TooltipValueFormatter,
 } from '../lib/utils/interactions';
-import { Scale, ScaleType } from '../lib/utils/scales/scales';
+import { Scale, ScaleType } from '../utils/scales/scales';
 import { DEFAULT_TOOLTIP_SNAP, DEFAULT_TOOLTIP_TYPE } from '../specs/settings';
 import { AnnotationDimensions, computeAnnotationDimensions, computeAnnotationTooltipState } from './annotation_utils';
 import { getCursorBandPosition, getCursorLinePosition, getTooltipPosition } from './crosshair_utils';
