+++ src/state/crosshair_utils.ordinal_snap.test.ts
@@ -1,7 +1,7 @@
 import { computeXScale } from '../lib/series/scales';
 import { BasicSeriesSpec } from '../lib/series/specs';
 import { getGroupId, getSpecId } from '../utils/ids';
-import { ScaleType } from '../lib/utils/scales/scales';
+import { ScaleType } from '../utils/scales/scales';
 import { getSnapPosition } from './crosshair_utils';
 import { computeSeriesDomains } from './utils';
