+++ src/state/test/interactions.test.ts
@@ -4,8 +4,8 @@ import { DataSeriesColorsValues } from '../../lib/series/series';
 import { BarSeriesSpec, BasicSeriesSpec, RectAnnotationSpec } from '../../lib/series/specs';
 import { getAnnotationId, getGroupId, getSpecId } from '../../utils/ids';
 import { TooltipType } from '../../lib/utils/interactions';
-import { ScaleContinuous } from '../../lib/utils/scales/scale_continuous';
-import { ScaleType } from '../../lib/utils/scales/scales';
+import { ScaleContinuous } from '../../utils/scales/scale_continuous';
+import { ScaleType } from '../../utils/scales/scales';
 import { ChartStore } from '../chart_state';
 import { computeSeriesDomains } from '../utils';
 import { ScaleBand } from '../../lib/utils/scales/scale_band';
