+++ src/utils/scales/scale_band.test.ts
